refactor command tests

+++ src/main/java/ua/in/denoming/sqlcmd/model/command/Delete.java
@@ -26,10 +26,16 @@ public class Delete implements Command {
         Validate.isTrue(canExecute(args));
 
         String tableName = args[0];
+        if (!databaseManager.isTableExists(tableName)) {
+            view.writeLine(String.format("Table with '%s' name not exists", tableName));
+            String tables = databaseManager.getTables().toString();
+            view.writeLine(tables);
+            return;
+        }
+
         String column = args[1];
         String searchValue = args[2];
         databaseManager.deleteData(tableName, column, searchValue);
-
         view.writeFormatLine("Data in table '%s' has deleted successfully", tableName);
     }
 }
